requirements.txt
asyncio
aiortc
requests
numpy
PySimpleGUI==4.60.5
pillow
ollama